model

+++ src/Infrastructure/Entitites/BillDetail.cs
@@ -5,11 +5,11 @@ namespace Infrastructure.Entitites
     public class BillDetail
     {
         [Key] 
-        public int Id { get; set; }
-        public int BillId { get; set; }
+        public long Id { get; set; }
+        public long BillId { get; set; }
         public Bill Bill { get; set; }
 
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public Product Product { get; set; }
         public decimal Quantity { get; set; }
         public decimal Price { get; set; }
